requirements.txt
affine==2.4.0
anyio==4.6.0
appnope==0.1.4
arrow==1.3.0
asttokens==2.4.1
attrs==24.2.0
bidict==0.23.1
binaryornot==0.4.4
blinker==1.9.0
branca==0.8.0
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.2
contextily==1.6.2
contourpy==1.3.0
cookiecutter==2.6.0
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
dnspython==2.7.0
eventlet==0.38.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
fiona==1.10.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
folium==0.17.0
fonttools==4.54.1
geographiclib==2.0
geopandas==0.14.4
geopy==2.4.1
greenlet==3.1.1
h11==0.14.0
humanize==4.11.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
ipyvue==1.11.1
ipyvuetify==1.10.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matlab==0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mercantile==1.2.1
Mesa==2.4.0
Mesa-Viz-Tornado==0.1.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
osmnx==1.9.4
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pymdown-extensions==10.11.2
pyogrio==0.10.0
pyparsing==3.1.4
pyproj==3.7.0
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-slugify==8.0.4
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rasterio==1.4.1
reacton==1.8.3
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rich-click==1.8.3
rpds-py==0.20.0
shapely==2.0.6
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
solara==1.39.0
solara-server==1.39.0
solara-ui==1.39.0
stack-data==0.6.3
starlette==0.39.2
text-unidecode==1.3
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
watchdog==5.0.3
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wsproto==1.2.0
xyzservices==2024.9.0
